requirements.txt
# Core video processing
opencv-python
Pillow
moviepy
numpy

# YouTube API
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-auth
# Scheduling and timezone
schedule
pytz

# AI integration with OpenAI
openai>=1.0.0
python-dotenv>=1.0.0

# Logging and utilities
python-dateutil

# Optional: For better performance
# ffmpeg-python==0.2.0  # Uncomment if you want programmatic ffmpeg control
